pdf chat bot/requirement.txt
streamlit
python-dotenv
google-generativeai
langchain-community
Pillow

